+++ packages/ui/Cancelable/Cancelable.js
@@ -46,8 +46,24 @@ const Cancelable = hoc(defaultConfig, (config, Wrapped) => {
 
 	invariant(onCancel, 'onCancel must be specified with Cancelable');
 
+	const onCancelIsString = typeof onCancel === 'string';
+	const onCancelIsFunction = typeof onCancel === 'function';
+	const dispatchCancelToConfig = function (props) {
+		if (onCancelIsString && typeof props[onCancel] === 'function') {
+			props[onCancel]();
+		} else if (onCancelIsFunction) {
+			return onCancel(props);
+		} else {
+			return true;
+		}
+	};
+
 	return class extends React.Component {
 		static displayName = 'Cancelable';
+
+		propTypes: {
+			onCancel: React.PropTypes.func
+		}
 
 		componentDidMount () {
 			if (modal) {
@@ -62,13 +78,11 @@ const Cancelable = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		cancel = () => {
-			if (typeof onCancel === 'string' && typeof this.props[onCancel] === 'function') {
-				this.props[onCancel]();
-			} else if (typeof onCancel === 'function') {
-				return onCancel(this.props);
-			} else {
-				return true;
+			if (this.props.onCancel) {
+				this.props.onCancel();
 			}
+
+			return dispatchCancelToConfig(this.props);
 		}
 
 		handleModalCancel = handle(
@@ -84,30 +98,33 @@ const Cancelable = hoc(defaultConfig, (config, Wrapped) => {
 			stopImmediate
 		)
 
-		renderWrapped () {
+		renderWrapped (props) {
 			return (
 				<Component onKeyUp={this.handleKeyUp}>
-					<Wrapped {...this.props} />
+					<Wrapped {...props} />
 				</Component>
 			);
 		}
 
-		renderUnwrapped () {
+		renderUnwrapped (props) {
 			return (
-				<Wrapped {...this.props} onKeyUp={this.handleKeyUp} />
+				<Wrapped {...props} onKeyUp={this.handleKeyUp} />
 			);
 		}
 
-		renderModal () {
+		renderModal (props) {
 			return (
-				<Wrapped {...this.props} />
+				<Wrapped {...props} />
 			);
 		}
 
 		render () {
-			return	modal && this.renderModal() ||
-					Component && this.renderWrapped() ||
-					this.renderUnwrapped();
+			const props = Object.assign({}, this.props);
+			delete props.onCancel;
+
+			return	modal && this.renderModal(props) ||
+					Component && this.renderWrapped(props) ||
+					this.renderUnwrapped(props);
 		}
 	};
 });