+++ packages/sampler/CHANGELOG.md
@@ -6,6 +6,9 @@ The following is a curated list of changes in the Enact project, newest changes 
 
 ### Added
 
+- Samples for `moonstone/DayPicker`, `moonstone/DatePicker`, `moonstone/ExpandableItem`, and
+	`moonstone/Image`.
+- Background selection tool
 - window.ReactPerf to the Sampler - available when switching to the "iframe" context in the dev tools
 
 ### Changed